requirements.txt
requests>=2.31.0
pycryptodome>=3.19.0
urllib3>=2.0.0

